# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Budget Splitter & Debt Graph API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: getGroups
      tags: [groups]
      summary: List groups for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupWithStats"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group by ID
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete group
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /groups/{groupId}/members:
    post:
      operationId: inviteMember
      tags: [groups]
      summary: Invite a member to a group
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberInput"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/members/{userId}:
    delete:
      operationId: removeMember
      tags: [groups]
      summary: Remove member from group
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /groups/{groupId}/settlements:
    get:
      operationId: getSettlements
      tags: [settlements]
      summary: List settlements in a group
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Settlements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Settlement"
    post:
      operationId: createSettlement
      tags: [settlements]
      summary: Record a settlement payment
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementInput"
      responses:
        "201":
          description: Settlement recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    GroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        currency:
          type: string
          default: USD
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        currency:
          type: string
    Group:
      type: object
      required: [id, name, currency, ownerId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        currency:
          type: string
        ownerId:
          type: integer
        createdAt:
          type: string
          format: date-time
    GroupWithStats:
      type: object
      required: [id, name, currency, ownerId, createdAt, memberCount, totalExpenses, myBalance]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        currency:
          type: string
        ownerId:
          type: integer
        createdAt:
          type: string
          format: date-time
        memberCount:
          type: integer
        totalExpenses:
          type: number
        myBalance:
          type: number
    GroupDetail:
      type: object
      required: [id, name, currency, ownerId, createdAt, members]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        currency:
          type: string
        ownerId:
          type: integer
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberWithBalance"
    Membership:
      type: object
      required: [id, groupId, userId, role, joinedAt]
      properties:
        id:
          type: integer
        groupId:
          type: integer
        userId:
          type: integer
        role:
          type: string
        joinedAt:
          type: string
          format: date-time
    MemberWithBalance:
      type: object
      required: [id, userId, role, joinedAt, user, balance]
      properties:
        id:
          type: integer
        userId:
          type: integer
        role:
          type: string
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        balance:
          type: number
    InviteMemberInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...
    SettlementInput:
      type: object
      required: [payerId, receiverId, amount]
      properties:
        payerId:
          type: integer
        receiverId:
          type: integer
        amount:
          type: number
        note:
          type: string
    Settlement:
      type: object
      required: [id, groupId, payerId, receiverId, amount, createdAt, payer, receiver]
      properties:
        id:
          type: integer
        groupId:
          type: integer
        payerId:
          type: integer
        receiverId:
          type: integer
        amount:
          type: number
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        payer:
          $ref: "#/components/schemas/User"
        receiver:
          $ref: "#/components/schemas/User"
...